Adjust working directory for epfmu simulation
The working simulation directory of an epfmu simulation is
defined as the current working directory + /instanceName.
The problem with this is when the instanceName is a fully qualified
path, (or something that looks like a fully qualified path),
in which case the fmu will be simulated in some unexpected
location.

This change uses only the cwd + /instanceName.filename() to handle the
case, when instance name is path like and absolute. In such a case we
use only the filename component of the instance name.